requirements.txt
streamlit==1.45.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
chromadb==1.0.8
openai==1.77.0
pymongo==4.12.1
python-dotenv==1.1.0
tiktoken==0.7.0
pydantic==2.6.4 